Fixed measure display crash
When the context of the spinner is a ContextThemeWrapper instead of an activity the current measureDisplayHeight will crash when attempting to cast the context to Activity. Changed the method to instead use the DisplayMetrics from resources to get the display height. This scenario can be replicated by placing the spinner inside a v7 Toolbar