Fix wrong position passed through listeners
Revert one of the last changes to pass again the right 'position' value as a parameter of listener callback method.
Due the nature of the spinner, a method call to `parent.getItemAtPosition` will return a wrong value, though, as the entries displayed by the pop-up list and the ones contained in the adapter are different: the selected item won't be displayed within the list.

To avoid getting a wrong value from the parent, I've deprecated the current listeners interface and created a brand new one that solves the problem. The sample project app has been update to show the new approach.

Also, added two suite of tests to test the selection logic and the listener.